mapping_script/pyproject.toml
[tool.poetry]
name = "mapping-script"
version = "0.1.0"
description = ""
authors = ["nvim <prozskov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.3"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
